# perf(circuitbreak): no allocation for MW